requirements.txt
vk-api==11.9.9
selenium==4.29.0
undetected-chromedriver==3.5.5
telethon==1.39.0
pytest==7.2.0
python-dotenv==1.0.1
pandas==2.2.3
datetime==5.4
requests==2.32.3
beautifulsoup4==4.12.3
emoji==2.14.1
aiohttp==3.10.11
torch==2.6.0  # ==2.6.0+cu126  Можно использовать версию с cuda
tensorflow==2.19.0
transformers==4.48.0
sentencepiece==0.2.0
mistralai==1.7.0
scikit-learn==1.6.1
optuna==4.3.0
psutil==7.0.0
md2pdf==1.0.1